+++ packages/temporal-utils/src/endOf.ts
@@ -1,4 +1,3 @@
-import type { Temporal } from 'temporal-spec'
 import {
   startOfHour,
   startOfMicrosecond,