+++ packages/temporal-utils/src/global-types.ts
@@ -1,0 +1,2 @@
+// Specifying with {} will strip this for JS build
+import {} from 'temporal-spec/global'